+++ code_editor/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log - @stonecrop/code-editor
 
+## 0.34.0
+
 ## 0.33.0
 
 ## 0.32.0