+++ nuxt/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Change Log - @stonecrop/nuxt
 
+## 0.34.0
+
+### Minor Changes
+
+- e9c340f: DocBuilder pages now render inside Desktop, and the `useDocBuilderRouteAdapter()` they use is auto-imported for any host page that does the same.
+
+### Patch Changes
+
+- Updated dependencies [e9c340f]
+- Updated dependencies [e9c340f]
+- Updated dependencies [e9c340f]
+- Updated dependencies [e9c340f]
+  - @stonecrop/desktop@0.34.0
+  - @stonecrop/themes@0.34.0
+  - @stonecrop/aform@0.34.0
+  - @stonecrop/atable@0.34.0
+  - @stonecrop/stonecrop@0.34.0
+  - @stonecrop/casl-middleware@0.34.0
+  - @stonecrop/code-editor@0.34.0
+  - @stonecrop/graphql-client@0.34.0
+  - @stonecrop/graphql-middleware@0.34.0
+  - @stonecrop/node-editor@0.34.0
+  - @stonecrop/nuxt-grafserv@0.34.0
+  - @stonecrop/schema@0.34.0
+
 ## 0.33.0
 
 ### Patch Changes